requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pandas==2.0.3
gunicorn==20.1.0
requests==2.31.0
deepdiff==6.7.1